refactor(cpu): split the two CPU generations into their own files
USE_FSMV2_CPU picks between two complete CPU paths, and both lived in one
840-line file with the gate buried among them. Answering "where does the
flag branch?" meant reading the whole file to be sure there was no second
fork.

The two paths now have their own files, so container_monitor.go holds the
types, the constructors, the single flag read, the single gate, and the
collectors that belong to neither generation. It is the only file that
mentions either path, and it mentions each exactly once, six lines apart.

  container_monitor.go  783 -> 407
  cpu_legacy.go         259  gopsutil, the cgroup reads, the old rules
  cpu_fsmv2.go          166  the worker's verdict and the seam

Memory, disk and HWID collection stay where they are. They belong to
neither generation, and moving them would triple the reviewable surface
of a commit whose only value is being obviously mechanical.

cgroupSnapshot also stays, although its only logic moves: it types a
field on ContainerMonitorService, and that struct cannot be read without
it.

This is a pure move plus six imports the compiler demands. Verified by
reconstruction rather than by inspection: container_monitor.go is
byte-equal to the original with two contiguous regions and those six
import lines removed, so there is no third edit anywhere in it. Each
moved declaration was compared to its original in order, without
sorting, because a sort-first comparison cannot see a reordering. Every
import in the new files was removed one at a time to confirm the build
breaks without it.

No test file changed. A pure move that needs one is not a pure move.
