+++ umh-core/pkg/service/container_monitor/container_monitor.go
@@ -19,7 +19,6 @@ import (
 	"crypto/rand"
 	"crypto/sha3"
 	"fmt"
-	"os"
 	"runtime"
 	"sync"
 	"time"
@@ -28,15 +27,10 @@ import (
 
 	"encoding/hex"
 
-	"github.com/shirou/gopsutil/v4/cpu"
 	"github.com/shirou/gopsutil/v4/disk"
 	"github.com/shirou/gopsutil/v4/mem"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/constants"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/cpuhealth"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/env"
-	fsmv2cpu "github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsmv2/cpu"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsmv2/fsmv2client"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsmv2/simple"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/logger"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/models"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/service/filesystem"
@@ -108,29 +102,6 @@ func NewContainerMonitorServiceWithPath(fs filesystem.Service, dataPath string) 
 		useFSMv2CPU:      useFSMv2CPU,
 		cpuUsageProvider: defaultCPUUsagePercent,
 	}
-}
-
-// defaultCPUUsagePercent reads the HOST's CPU usage through gopsutil: the first
-// element of the non-per-CPU percentage slice.
-//
-// This is the legacy CPU path's only usage source and it runs ONLY when
-// USE_FSMV2_CPU is off.
-func defaultCPUUsagePercent(ctx context.Context) (float64, error) {
-	usagePercentages, err := cpu.PercentWithContext(ctx, 0, false)
-	if err != nil {
-		return 0, err
-	}
-
-	if len(usagePercentages) > 0 {
-		return usagePercentages[0], nil
-	}
-
-	return 0, nil
-}
-
-// setCPUUsageProvider overrides the CPU usage source - used for testing only.
-func (c *ContainerMonitorService) setCPUUsageProvider(fn func(ctx context.Context) (float64, error)) {
-	c.cpuUsageProvider = fn
 }
 
 // GetFilesystemService returns the filesystem service - used for testing only.
@@ -259,353 +230,6 @@ func (c *ContainerMonitorService) GetHealth(ctx context.Context) (*models.Health
 	}
 
 	return health, nil
-}
-
-// cpuWorkerMaxAge is how old the fsmv2 CPU worker's observation may be and still
-// count as Fresh for the seam. It is 3x the worker's 1s poll interval
-// (pollInterval in pkg/fsmv2/cpu), so one slow or missed poll cannot flip the
-// instance to degraded.
-const cpuWorkerMaxAge = 3 * time.Second
-
-// judgeLegacyCPUUsage degrades the instance above CPUHighThresholdPercent of the
-// cores it may use, and only collectCPULegacy calls it, so it never runs under
-// USE_FSMV2_CPU.
-//
-// getCPUMetrics already degrades on that same threshold, so this changes an
-// outcome only where the two compute different core counts: a quota below 0.1
-// cores, which getCPUMetrics clamps and this does not. See ENG-5384.
-func judgeLegacyCPUUsage(status *ServiceInfo, cpuStat *models.CPU) {
-	effectiveCores := cpuStat.CgroupCores
-	if effectiveCores <= 0 && cpuStat.CoreCount != nil {
-		// Fall back to host cores if cgroup info unavailable
-		effectiveCores = float64(*cpuStat.CoreCount)
-	}
-
-	// A nil TotalUsageMCpu records a tick nothing measured, so there is no
-	// percentage to compute.
-	if effectiveCores > 0 && cpuStat.TotalUsageMCpu != nil {
-		cpuPercent := (*cpuStat.TotalUsageMCpu / 1000.0) / effectiveCores * 100.0
-
-		if cpuPercent > constants.CPUHighThresholdPercent {
-			status.CPUHealth = models.Degraded
-			status.OverallHealth = models.Degraded
-		}
-	}
-}
-
-// collectCPULegacy runs the pre-fsmv2 CPU path.
-func (c *ContainerMonitorService) collectCPULegacy(ctx context.Context, status *ServiceInfo) error {
-	cpuStat, err := c.getCPUMetrics(ctx)
-	if err != nil {
-		return err
-	}
-
-	status.CPU = cpuStat
-	judgeLegacyCPUUsage(status, cpuStat)
-
-	return nil
-}
-
-// collectCPUFromWorker builds the whole CPU record from the fsmv2 CPU worker's
-// last observation. The legacy fields stay empty on purpose: old and new
-// reporting stay cleanly separated, so nothing here re-derives a legacy-named
-// number from worker data. models.CPU says what each generation carries.
-func (c *ContainerMonitorService) collectCPUFromWorker(ctx context.Context) *models.CPU {
-	health, cpuHealth := c.readWorkerCPUHealth(ctx)
-
-	return &models.CPU{Health: health, CPUHealth: cpuHealth}
-}
-
-// readWorkerCPUHealth reads the fsmv2 CPU worker's observation and maps it to a
-// models.Health. It always returns a health, and the protocol converter's
-// IsResourceLimited reads that message as its bridge-block reason.
-func (c *ContainerMonitorService) readWorkerCPUHealth(ctx context.Context) (health *models.Health, cpuHealth *models.CPUHealth) {
-	client := fsmv2client.GetClient()
-	// Fallback for a misconfiguration: USE_FSMV2_CPU is on but nothing published
-	// a client, so the fsmv2 supervisor never started (or has not yet).
-	if client == nil {
-		message := c.cpuSeamClientUnavailableMessage()
-
-		c.cpuWorkerWarnOnce.Do(func() {
-			c.logger.Warn(message)
-		})
-
-		return &models.Health{
-			Message:       message,
-			ObservedState: models.Degraded.String(),
-			DesiredState:  models.Active.String(),
-			Category:      models.Degraded,
-		}, nil
-	}
-
-	// Get the latest poll result from the worker.
-	workerStatus, freshness, err := fsmv2client.GetFresh[simple.Status[fsmv2cpu.CPUStatus]](ctx, client, fsmv2cpu.Ref, cpuWorkerMaxAge)
-	if err != nil {
-		// GetFresh returns Unknown here: the read failure prevented it from being
-		// classified, so it cannot be called healthy. Fail closed with the
-		// verbatim store error as the message.
-		return &models.Health{
-			Message:       fmt.Sprintf("CPU worker observation could not be read: %v", err),
-			ObservedState: models.Degraded.String(),
-			DesiredState:  models.Active.String(),
-			Category:      models.Degraded,
-		}, nil
-	}
-
-	// If it is not fresh, handle these cases here.
-	if freshness != fsmv2client.Fresh {
-		message := "CPU worker observation could not be classified; no measurement to judge"
-
-		switch freshness {
-		case fsmv2client.Stale:
-			message = fmt.Sprintf("CPU worker observation is stale (older than %s); cannot trust the verdict it carries", cpuWorkerMaxAge)
-		case fsmv2client.NeverObserved:
-			message = "CPU worker has never observed; no measurement to judge"
-		case fsmv2client.Unregistered:
-			message = "CPU worker is not registered with the fsmv2 runtime; no measurement to judge"
-		}
-
-		return &models.Health{
-			Message:       message,
-			ObservedState: models.Degraded.String(),
-			DesiredState:  models.Active.String(),
-			Category:      models.Degraded,
-		}, nil
-	}
-
-	// Degraded without a degraded verdict means the poll failed, not that the box
-	// is degraded.
-	if workerStatus.Degraded && workerStatus.Result.Verdict.State != cpuhealth.StateDegraded {
-		return &models.Health{
-			Message:       workerStatus.Reason,
-			ObservedState: models.Degraded.String(),
-			DesiredState:  models.Active.String(),
-			Category:      models.Degraded,
-		}, nil
-	}
-
-	// A Fresh observation carries the developer's judgement in Result. The
-	// switch maps only the two spelled-out states, so a rename of either
-	// fails the seam tests too.
-	switch workerStatus.Result.Verdict.State {
-	case cpuhealth.StateHealthy:
-		return &models.Health{
-				Message:       workerStatus.Result.Message,
-				ObservedState: models.Active.String(),
-				DesiredState:  models.Active.String(),
-				Category:      models.Active,
-			}, &models.CPUHealth{
-				Verdict: workerStatus.Result.Verdict,
-				Details: workerStatus.Result.Details,
-			}
-	case cpuhealth.StateDegraded:
-		return &models.Health{
-				Message:       workerStatus.Result.Message,
-				ObservedState: models.Degraded.String(),
-				DesiredState:  models.Active.String(),
-				Category:      models.Degraded,
-			}, &models.CPUHealth{
-				Verdict: workerStatus.Result.Verdict,
-				Details: workerStatus.Result.Details,
-			}
-	default:
-		// Empty result verdict AND Degraded == false is a genuine "no
-		// determination" — a successful poll produced no verdict. There is no
-		// second opinion to defer to, so say so rather than read it as healthy.
-		return &models.Health{
-			Message:       "CPU worker produced no verdict for its last observation",
-			ObservedState: models.Degraded.String(),
-			DesiredState:  models.Active.String(),
-			Category:      models.Degraded,
-		}, nil
-	}
-}
-
-// cpuSeamClientUnavailableMessage says which prerequisite is missing when
-// USE_FSMV2_CPU is on but no CPU worker client was published, or that the
-// supervisor may still be starting.
-func (c *ContainerMonitorService) cpuSeamClientUnavailableMessage() string {
-	transportOn, _ := env.GetAsBool("USE_FSMV2_TRANSPORT", false, true)
-	if !transportOn {
-		return "USE_FSMV2_CPU is enabled but USE_FSMV2_TRANSPORT is off, so the fsmv2 supervisor never runs and no CPU worker client is published; no CPU measurement is available"
-	}
-
-	if os.Getenv("API_URL") == "" || os.Getenv("AUTH_TOKEN") == "" {
-		return "USE_FSMV2_CPU is enabled but API_URL or AUTH_TOKEN is unset, so the fsmv2 supervisor never runs and no CPU worker client is published; no CPU measurement is available"
-	}
-
-	return "USE_FSMV2_CPU is enabled but no fsmv2 client is reachable yet (the fsmv2 supervisor may still be starting); no CPU measurement is available"
-}
-
-// getCPUMetrics collects CPU metrics using gopsutil.
-// By default, this retrieves host-level usage unless gopsutil is configured
-// to read from container cgroup data. See notes below for cgroup-limited usage.
-func (c *ContainerMonitorService) getCPUMetrics(ctx context.Context) (*models.CPU, error) {
-	usageMCores, coreCount, usagePercent, err := c.getRawCPUMetrics(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	// Get cgroup info for throttling and limits
-	cgroupInfo, cgroupErr := c.getCgroupCPUInfo(ctx)
-	if ctx.Err() != nil {
-		return nil, ctx.Err()
-	}
-
-	// Default to Active health
-	category := models.Active
-	message := "CPU utilization normal"
-
-	// Compute windowed throttle ratio; skip entirely on cgroup read failure to preserve wasThrottled state
-	var (
-		windowedRatio float64
-		isThrottled   bool
-	)
-	if cgroupErr == nil && cgroupInfo != nil {
-		windowedRatio, isThrottled = c.updateThrottleWindow(cgroupInfo)
-		cgroupInfo.ThrottleRatio = windowedRatio
-		cgroupInfo.IsThrottled = isThrottled
-
-		if isThrottled && !c.wasThrottled {
-			c.logger.Warnf("CPU throttling detected: %.1f%% of periods throttled", cgroupInfo.ThrottleRatio*100)
-		}
-
-		c.wasThrottled = isThrottled
-	}
-
-	switch {
-	case usagePercent >= constants.CPUHighThresholdPercent || isThrottled:
-		category = models.Degraded
-
-		if isThrottled && cgroupInfo != nil {
-			message = fmt.Sprintf("CPU throttled (%.1f%% periods throttled)", cgroupInfo.ThrottleRatio*100)
-		} else {
-			message = "CPU utilization critical"
-		}
-	case usagePercent >= constants.CPUMediumThresholdPercent:
-		message = "CPU utilization warning"
-	}
-
-	cpuStat := &models.CPU{
-		Health: &models.Health{
-			Message:       message,
-			ObservedState: category.String(),
-			DesiredState:  models.Active.String(),
-			Category:      category,
-		},
-		TotalUsageMCpu: &usageMCores,
-		CoreCount:      &coreCount,
-	}
-
-	// Add cgroup info if available
-	if cgroupErr == nil {
-		cpuStat.CgroupCores = cgroupInfo.QuotaCores
-		cpuStat.ThrottleRatio = cgroupInfo.ThrottleRatio
-		cpuStat.IsThrottled = cgroupInfo.IsThrottled
-	}
-
-	return cpuStat, nil
-}
-
-// updateThrottleWindow appends a cgroup snapshot and computes the throttle ratio
-// over a sliding window defined by constants.CPUThrottleWindow.
-// Returns (0.0, false) when there is insufficient data, nil input, or counter reset.
-func (c *ContainerMonitorService) updateThrottleWindow(cgroupInfo *CPUCgroupInfo) (ratio float64, isThrottled bool) {
-	// Guard: nil input or zero periods (cpu.stat unreadable)
-	if cgroupInfo == nil || cgroupInfo.NrPeriods <= 0 {
-		return 0.0, false
-	}
-
-	now := time.Now()
-
-	// Detect counter reset: if new counters are lower than the newest snapshot,
-	// the cgroup was recreated (pod rescheduled). Clear buffer and start fresh.
-	if len(c.throttleSnapshots) > 0 {
-		newest := c.throttleSnapshots[len(c.throttleSnapshots)-1]
-		if cgroupInfo.NrPeriods < newest.nrPeriods || cgroupInfo.NrThrottled < newest.nrThrottled {
-			c.throttleSnapshots = nil
-		}
-	}
-
-	// Append current snapshot
-	c.throttleSnapshots = append(c.throttleSnapshots, cgroupSnapshot{
-		timestamp:   now,
-		nrPeriods:   cgroupInfo.NrPeriods,
-		nrThrottled: cgroupInfo.NrThrottled,
-	})
-
-	// Prune entries older than the window
-	cutoff := now.Add(-constants.CPUThrottleWindow)
-
-	pruneIdx := 0
-	for pruneIdx < len(c.throttleSnapshots) && c.throttleSnapshots[pruneIdx].timestamp.Before(cutoff) {
-		pruneIdx++
-	}
-
-	if pruneIdx > 0 {
-		c.throttleSnapshots = c.throttleSnapshots[pruneIdx:]
-	}
-
-	// Need at least 2 snapshots for a delta
-	if len(c.throttleSnapshots) < 2 {
-		return 0.0, false
-	}
-
-	// Compute delta between newest and oldest snapshot in window
-	oldest := c.throttleSnapshots[0]
-	current := c.throttleSnapshots[len(c.throttleSnapshots)-1]
-
-	deltaPeriods := current.nrPeriods - oldest.nrPeriods
-	deltaThrottled := current.nrThrottled - oldest.nrThrottled
-
-	if deltaPeriods <= 0 {
-		return 0.0, false
-	}
-
-	ratio = float64(deltaThrottled) / float64(deltaPeriods)
-	isThrottled = ratio > constants.CPUThrottleRatioThreshold
-
-	return ratio, isThrottled
-}
-
-func (c *ContainerMonitorService) getRawCPUMetrics(ctx context.Context) (usageMCores float64, coreCount int, usagePercent float64, err error) {
-	// Try to get cgroup info first for accurate container limits
-	cgroupInfo, cgroupErr := c.getCgroupCPUInfo(ctx)
-	if ctx.Err() != nil {
-		return 0, 0, 0, ctx.Err()
-	}
-
-	// Get actual CPU usage through the injectable source (gopsutil by default).
-	provider := c.cpuUsageProvider
-	if provider == nil {
-		provider = defaultCPUUsagePercent
-	}
-	usagePercent, err = provider(ctx)
-	if err != nil {
-		return 0, 0, 0, err
-	}
-
-	// Determine effective core count (keep as float64 to preserve fractional quotas)
-	// Use cgroup limit if available, otherwise fall back to host CPU count
-	effectiveCores := float64(runtime.NumCPU())
-	if cgroupErr == nil && cgroupInfo.QuotaCores > 0 {
-		// Use cgroup limit for more accurate mCPU calculation
-		// QuotaCores can be fractional (e.g., 0.5 for 500m, 1.5 for 1500m)
-		effectiveCores = cgroupInfo.QuotaCores
-		// Use a small minimum to avoid divide-by-zero, but preserve fractional limits
-		if effectiveCores < 0.1 {
-			effectiveCores = 0.1
-		}
-	}
-
-	coreCount = runtime.NumCPU() // Always report host cores for compatibility
-
-	// Convert usage percent to mCPU based on effective cores
-	// This gives us a more accurate representation when cgroups limit CPU
-	usageCores := (usagePercent / 100.0) * effectiveCores
-	usageMCores = usageCores * 1000
-
-	return usageMCores, coreCount, usagePercent, nil
 }
 
 // getMemoryMetrics collects memory metrics, preferring cgroup values when available.
